implementing clang tidy suggestions

+++ src/Entity.cpp
@@ -63,7 +63,7 @@ void Entity::Draw()
 void Entity::DrawHealthBar() const
 {
     pdcpp::GlobalPlaydateAPI::get()->graphics->drawRect(position.x-5, position.y - 10, 25, 4, kColorWhite);
-    pdcpp::GlobalPlaydateAPI::get()->graphics->fillRect(position.x-5, position.y - 10, 25 * (hp / maxHP), 4, kColorWhite);
+    pdcpp::GlobalPlaydateAPI::get()->graphics->fillRect(position.x-5, position.y - 10, 25 * static_cast<int>(hp / maxHP), 4, kColorWhite);
 }
 
 void Entity::Damage(float damage)
